Moved basic parallelization code to multithreading.hpp
