+++ src/simulator/multithreading.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "op_graph.hpp"
+
+namespace CoreIR {
+
+  void balancedComponentsParallel(NGraph& gr);
+}
